+++ lib/command/prebuild.rb
@@ -8,16 +8,23 @@ module Pod
         self.arguments = [CLAide::Argument.new("CACHE-BRANCH", false)]
         def self.options
           [
+            ["--config", "Config (Debug, Test...) to prebuild"],
             ["--push", "Push cache to repo upon completion"],
             ["--all", "Prebuild all binary pods regardless of cache validation"],
             ["--targets", "Targets to prebuild. Use comma (,) to specify a list of targets"]
-          ]
+          ].concat(super)
         end
 
         def initialize(argv)
           super
-          @prebuild_all_pods = argv.flag?("all")
-          @prebuild_targets = argv.option("targets", "").split(",")
+          prebuild_all_pods = argv.flag?("all")
+          prebuild_targets = argv.option("targets", "").split(",")
+          update_cli_config(
+            :prebuild_job => true,
+            :prebuild_all_pods => prebuild_all_pods,
+            :prebuild_config => argv.option("config")
+          )
+          update_cli_config(:prebuild_targets => prebuild_targets) unless prebuild_all_pods
           @prebuilder = PodPrebuild::CachePrebuilder.new(
             config: prebuild_config,
             cache_branch: argv.shift_argument || "master",
@@ -26,9 +33,6 @@ module Pod
         end
 
         def run
-          PodPrebuild.config.cli_config[:prebuild_job] = true
-          PodPrebuild.config.cli_config[:prebuild_all_pods] = @prebuild_all_pods
-          PodPrebuild.config.cli_config[:prebuild_targets] = @prebuild_targets unless @prebuild_all_pods
           @prebuilder.run
         end
       end