+++ spec/command/config_spec.rb
@@ -1,0 +1,3 @@
+describe "Pod::Config" do
+  # TODO (thuyen): Write tests
+end